fix(package): move auth/session types to validation package
